pyproject.toml
[tool.poetry]
name = "aoc2022"
version = "0.1.0.dev"
description = ""
authors = ["Samuel Suter <smurtlesam@gmail.com>"]

[tool.poetry.scripts]
aoc = "aoc2022.common.cli:main"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
click = "^8.1.3"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
numpy = "^1.23.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
yapf = "^0.32.0"
mypy = "^0.982"
isort = "^5.10.1"
toml = "^0.10.2"
autoflake8 = "^0.4.0"
when-changed = {url = "https://github.com/joh/when-changed/archive/master.zip"}
types-requests = "^2.28.11.5"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 100

[tool.yapf]
based_on_style = "pep8"
column_limit = 100
BLANK_LINE_BEFORE_NESTED_CLASS_OR_DEF = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
